use generic useDebouncedValue hook for text inputs
