+++ x-pack/plugins/lens/public/indexpattern_datasource/dimension_panel/dimension_editor.tsx
@@ -7,7 +7,7 @@
 
 import './dimension_editor.scss';
 import _ from 'lodash';
-import React, { useState, useMemo, useEffect, useRef } from 'react';
+import React, { useState, useMemo } from 'react';
 import { i18n } from '@kbn/i18n';
 import {
   EuiListGroup,
@@ -44,6 +44,7 @@ import { ReferenceEditor } from './reference_editor';
 import { setTimeScaling, TimeScaling } from './time_scaling';
 import { defaultFilter, Filtering, setFilter } from './filtering';
 import { AdvancedOptions } from './advanced_options';
+import { useDebouncedValue } from '../../shared_components';
 
 const operationPanels = getOperationDisplay();
 
@@ -59,34 +60,7 @@ export interface DimensionEditorProps extends IndexPatternDimensionEditorProps {
  * if no debounced changes are in flight because the user is currently typing into the input.
  */
 const LabelInput = ({ value, onChange }: { value: string; onChange: (value: string) => void }) => {
-  const [inputValue, setInputValue] = useState(value);
-  const unflushedChanges = useRef(false);
-
-  // Save the initial value
-  const initialValue = useRef(value);
-
-  const onChangeDebounced = useMemo(() => {
-    const callback = _.debounce((val: string) => {
-      onChange(val);
-      unflushedChanges.current = false;
-    }, 256);
-    return (val: string) => {
-      unflushedChanges.current = true;
-      callback(val);
-    };
-  }, [onChange]);
-
-  useEffect(() => {
-    if (!unflushedChanges.current && value !== inputValue) {
-      setInputValue(value);
-    }
-  }, [value, inputValue]);
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const val = String(e.target.value);
-    setInputValue(val);
-    onChangeDebounced(val || initialValue.current);
-  };
+  const { inputValue, handleInputChange, initialValue } = useDebouncedValue({ onChange, value });
 
   return (
     <EuiFormRow
@@ -101,8 +75,10 @@ const LabelInput = ({ value, onChange }: { value: string; onChange: (value: stri
         compressed
         data-test-subj="indexPattern-label-edit"
         value={inputValue}
-        onChange={handleInputChange}
-        placeholder={initialValue.current}
+        onChange={(e) => {
+          handleInputChange(e.target.value);
+        }}
+        placeholder={initialValue}
       />
     </EuiFormRow>
   );
@@ -520,32 +496,41 @@ export function DimensionEditor(props: DimensionEditorProps) {
       {!currentFieldIsInvalid && (
         <div className="lnsIndexPatternDimensionEditor__section">
           {!incompleteInfo && selectedColumn && (
-            <LabelInput
-              value={selectedColumn.label}
-              onChange={(value) => {
-                setState(
-                  mergeLayer({
-                    state,
-                    layerId,
-                    newLayer: {
-                      columns: {
-                        ...state.layers[layerId].columns,
-                        [columnId]: {
-                          ...selectedColumn,
-                          label: value,
-                          customLabel:
-                            operationDefinitionMap[selectedColumn.operationType].getDefaultLabel(
-                              selectedColumn,
-                              state.indexPatterns[state.layers[layerId].indexPatternId],
-                              state.layers[layerId].columns
-                            ) !== value,
+            <EuiFormRow
+              label={i18n.translate('xpack.lens.indexPattern.columnLabel', {
+                defaultMessage: 'Display name',
+                description: 'Display name of a column of data',
+              })}
+              display="columnCompressed"
+              fullWidth
+            >
+              <LabelInput
+                value={selectedColumn.label}
+                onChange={(value) => {
+                  setState(
+                    mergeLayer({
+                      state,
+                      layerId,
+                      newLayer: {
+                        columns: {
+                          ...state.layers[layerId].columns,
+                          [columnId]: {
+                            ...selectedColumn,
+                            label: value,
+                            customLabel:
+                              operationDefinitionMap[selectedColumn.operationType].getDefaultLabel(
+                                selectedColumn,
+                                state.indexPatterns[state.layers[layerId].indexPatternId],
+                                state.layers[layerId].columns
+                              ) !== value,
+                          },
                         },
                       },
-                    },
-                  })
-                );
-              }}
-            />
+                    })
+                  );
+                }}
+              />
+            </EuiFormRow>
           )}
 
           {!incompleteInfo && !hideGrouping && (
