+++ x-pack/plugins/lens/public/shared_components/index.ts
@@ -9,3 +9,4 @@ export * from './empty_placeholder';
 export { ToolbarPopoverProps, ToolbarPopover } from './toolbar_popover';
 export { LegendSettingsPopover } from './legend_settings_popover';
 export { PalettePicker } from './palette_picker';
+export { useDebouncedValue } from './debounced_value';
